requirements.txt
streamlit
manim
google-genai
ffmpeg-python
pydantic
python-dotenv
